# src/main/resources/OpenApi.yaml
openapi: 3.0.3
info:
  title: API de proyecto web para reserva de alojamientos
  description: |-
    Api rest para gestion de alojamientos, usuarios y reservas.
    
    El proyecto es desarrola por la parte de Bakend en spring boot y uso angular para el fronted,
      además del uso de Maria DB como base de datos relacional.
    
    Enlaces de utilidad:
      -[Enlace de GitHub](https://github.com/Elkinlol/Proyecto-final)
      
    Contacto: elkin.bermudezg@uqvirtual.edu.co
  version: 1.1
servers:
  - url: http/:localhost:8082/api
    description: servidor local
...
paths:
  /Auth:
    post:
      tags: [Auth]
      summary: Creacion de un nuevo usuario
      operationId: registrarUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: Usuario registrado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Datos no validos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Correo ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /Auth/login:
    post:
      tags: [Auth]
      summary: Inicio de sesion -- JWT
      security:
        - BearerAuth: []
      operationId: iniciarSesion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginRequest'
            example:
              email: "usuario@ejemplo.com"
              password: "ContraseñaSegura123"
      responses:
        '200':
            description: Se ha iniciado sesión con exito
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/LoginResponse'
        '400':
            description: Faltan datos requridos (rellenar todos los campos)
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '401':
            description: credenciales invalidas (email o contraseña incorrecta)
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '404':
            description: Usuario no encontrado
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '500':
            description: error del servidor
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
...
  /auth/password:
    post:
      tags: [Auth]
      summary: Solicitar código de recuperación de contraseña
      security:
        - BearerAuth: []
      description: Envía un código de recuperación al correo electrónico del usuario. El código expira en 15 minutos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryResponse"
      responses:
        "200":
          description: Código de recuperación enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Se ha enviado un código de recuperación a su correo electrónico."
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Error del sevidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change/password:
    put:
      tags: [Auth]
      summary: Se cambia la contraseña
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        '200':
          description: contraseña cambiada de manera exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

        '400':
          description: Codigo invalido o expirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        '500':
          description: error del sevidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /listings:
    post:
      tags: [Listings]
      summary: crear un nuevo alojamiento
      operationId: createListing
      security:
        - bearerAuth: [ ]
      requestBody:
        content:
          schema:
            $ref: "#/components/schemas/CreateListingRequest"
      responses:
        '201':
          description: Alojamiento creado exitosamente
          content:
            schema:
              $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          "401":
            description: No autorizado (JWT inválido o ausente)
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"
          "403":
            description: No tiene permisos para realizar la accion
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"
          "500":
            description: error del servidor
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"
  /listings/{id}:
    put:
      summary: Editar un alojamiento existente
      tags:  [Listings]
      operationId: editListing
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del alojamiento a editar
          schema:
            type: integer
            example: 101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingRequest"
      responses:
        "200":
          description: Alojamiento actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: El usuario no es propietario del alojamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Eliminar un alojamiento (soft delete)
      tags: [Listings]
      operationId: deleteListing
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del alojamiento a eliminar
          schema:
            type: integer
            example: 101
      responses:
        "200":
          description: Alojamiento marcado como eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteListingResponse"
        "400":
          description: El alojamiento no puede eliminarse (reservas futuras)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: El usuario no es propietario del alojamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Listings]
      summary: Obtener los alojamientos -- ver sus detalles
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del alojamiento a eliminar
          schema:
            type: integer
            example: 101
      responses:
        '200':
          description: Se obtiene el alojamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        '401':
          description: No tiene autorizacion
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: No permisos para realizar esta opcion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
                
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /listings/search:
    get:
      summary: Buscar alojamientos
      tags: [ Listings ]
      operationId: searchListings
      security:
        - bearerAuth: [ ]
      parameters:
        - name: precio
          in: query
          required: false
          description: precio del alojamiento
          schema:
            type: number
            example: 2000
        - name: ciudad
          in: query
          required: false
          description: Ciudad donde se encuentra el alojamiento
          schema:
            type: string
            example: "Berlin"
        - name: fechaInicio
          in: query
          required: false
          description: Fecha de inicio de la búsqueda (check-in)
          schema:
            type: string
            format: date
            example: "2025-12-15"
        - name: fechaFin
          in: query
          required: false
          description: Fecha de fin de la búsqueda (check-out)
          schema:
            type: string
            format: date
            example: "2025-12-20"
        - name: servicios
          in: query
          required: false
          description: lista de servicios separado por coma (wifi, piscina, parqueadero)
          schema:
            type: array
            items:
              type: string
              example: "wifi"
            style: form
            explode: false
      responses:
        "200":
          description: Resultados de la búsqueda
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 3
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/ListingResponse"
        "400":
          description: Parámetros de búsqueda inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No se encontraron alojamientos que coincidan con la búsqueda
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /comments/listing/{id}:
    post:
      summary: Crear una nueva review desde huesped
      tags: [Comments]
      operationId: createComment
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Id de la reserva que se va a comentar
          schema:
            type: string
            example: 2314
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        '201':
          description: El comentario fue creado con exito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        '400':
          description: El comentario no fue creado por falta del mensaje
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: No se pudo crear el comentario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: No se pudo crear el comentario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /listings/{id}/promotion:
    get:
      summary: Listar las promociones
      tags:
        - Promotion
      operationId: listListindByIdPromotion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 9513
      responses:
        '200':
          description: Promociones enlistadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"      
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Promociones enlistadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Crear una nueva promocion
      tags:
        - Promotion
      operationId: createPromotion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          requiered: True
          schema:
            type: String
            Example: 4120
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromotionRequest"
      responses:
        '201':
          description: Promocion creada exitosamente
          content:
            application/json:
              $ref: "#/components/schemas/PromotionResponse"
        '400':
          description: Datos inválidos (fechas o porcentaje incorrecto)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: No tiene permisos para gestionar este alojamiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Error de servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    CreatePromotionRequest:
      type: object
      properties:
        descuento:
          type: number
          description: Porcentaje de descuento (1-100)
          example: 15
        fechaInicio:
            type: string
            format: date
            example: "2025-12-01"
        fechaFin:
            type: string
            format: date
            example: "2025-12-31"
        required:
          - descuento
          - fechaInicio
          - fechaFin
    PromotionResponse:
      type: object
      properties:
        id:
          type: integer
          example: 55
        alojamientoId:
          type: integer
          example: 101
        descuento:
          type: number
          example: 15
        fechaInicio:
          type: string
          format: date
          example: "2025-12-01"
        fechaFin:
          type: string
          format: date
          example: "2025-12-31"
        estado:
          type: string
          enum: [ Activa, Expirada ]
          example: Activa
...
    CreateCommentRequest:
      type: object
      properties:
        calificacion:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comentario:
          type: string
          maxLength: 500
          example: "El lugar estaba impecable y el anfitrión fue muy amable."
      required:
        - calificacion
        - comentario
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          example: 789
        reservaId:
          type: integer
          example: 1234
        usuarioId:
          type: integer
          example: 45
        calificacion:
          type: integer
          example: 5
        comentario:
          type: string
          example: "El lugar estaba impecable y el anfitrión fue muy amable."
        fechaCreacion:
          type: string
          format: date-time
          example: "2025-09-07T15:45:00Z"
...
    ListingResponse:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ubicacion:
          type: object
        PriceForNight:
          type: number
        Capacity:
          type: integer
        imagenes:
          type: array
          items:
            type: string
            format: uri
        anfitrionId:
          type: integer
          example: 55
        fechaPublicacion:
          type: string
          format: date-time
          example: "2025-09-07T13:00:00Z"
      required:
        - title
        - ubicacion
        - priceForNight
        - description
        - capacity
    DeleteListingResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "El alojamiento con id 123 fue eliminado"
    UpdateListingRequest:
      type: object
      properties:
        title:
          type: string
          example: Pent house
        descition:
          type: string
          example: Las mejores vistas de medellin
        ubicacion:
          type: object
          example: {direccion: calle 14, ciudad: medellin, latitud: 100, longitud: 200}
        priceForNight:
          type: number
          format: float
          example: 200
        Capacity:
          type: integer
          example: 40
        imagenes:
          type: array
          items:
            type: string
            format: uri
    CreateListingRequest:
      type: object
      required:
        -ubicacion
        -price
        -capacity
        -title
        -desciptionListing
        -imagenes

      properties:
        ubicacion:
          type: object
        price:
          type: integer
          description: precio por noche del alojamiento
          example: 20000
        title:
          type: string
          description: Nombre del alojamiento
          example: La finquita
        descriptionListing:
          type: string
          description: descripcion detallada del alojamiento
          example: Alojamiento a los alrededores de la montaña con excelente clima
        capacity:
          type: integer
          description: capacidad maxima del alojamiento
          example: 20
        service:
          type: string
          description: que servicios vienen con el alojamiento
          example: Wifi, cocina, etc
        imagenes:
          type: array
          items:
            type: string
            format: uri
    ChangePasswordRequest:
      type: object
      required:
        - new password
        - codigo
      properties:
        new password:
          type: string
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).*$"
          description: La contraseña del usuario (debe contener al menos un dígito, una letra minúscula y una letra mayúscula)
        codigo:
          type: string
          example: 9802
    RecoveryResponse:
      type: object
      required:
          - email
      properties:
        email:
            type: string
            format: email
            example: "carlos.perez@mail.com"
...
    UserRegister:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).*$"
          description: La contraseña del usuario (debe contener al menos un dígito, una letra minúscula y una letra mayúscula)
        fullName:
          type: string
          maxLength: 100
        dateBirth:
          type: string
          format: date
          example: "1990-01-01"
        role:
          type: string
          enum:
            - user
            - admin
          default: user
      required:
        - fullName
        - email
        - password
    loginRequest:
      type: object
      required: true
        -password
        -email
      properties:
        email:
          type: string
          format: email
          example: "usuario@ejemplo.com"
        password:
          type: string
          format: password
          example: "ContraseñaSegura123"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT de autenticación
          user:
            type: object
            properties:
              id:
                type: integer
              nombre:
                type: string
              email:
                type: string
                format: email
              rol:
                type: string
                enum: [ usuario, anfitrion]
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "Contraseña actualizada exitosamente"
      required:
        - message
    UserResponse:
      type: object
      properties:
        numberPhone:
          type: string
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        dateBirth:
          type: string
          format: date
        role:
          type: string
      required:
        - id
        - email
        - fullName
        - numberPhone
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT